+++ docs/versionhistory.rst
@@ -11,6 +11,8 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
   ``@staticmethod`` (PR by jacobpbrugh)
 - Changed ``check_type()`` to return the passed value, so it can be used (to an extent)
   in place of ``typing.cast()``, but with run-time type checking
+- Replaced custom implementation of ``is_typeddict()`` with the implementation from
+  ``typing_extensions`` v4.1.0
 - Fixed ``TypeError`` when checking against ``TypedDict`` when the value has mixed types
   among the extra keys (PR by biolds)
 - Fixed the configured ``forward_ref_policy`` not being used in